Task 1 comlete

+++ Task 1/src/Phone.java
@@ -1,11 +1,11 @@
 public class Phone {
-    int phoneNumber;
+    long phoneNumber;
     String phoneModel;
     double phoneWeight;
     static String madeIn="China";
     Phone(){
     }
-    Phone(int phoneNumber,String phoneModel, double phoneWeight,String madeIn){
+    Phone(long phoneNumber,String phoneModel, double phoneWeight,String madeIn){
         this.phoneNumber=phoneNumber;
         this.phoneModel=phoneModel;
         this.phoneWeight=phoneWeight;
@@ -14,7 +14,7 @@ public class Phone {
     void receiveCall(String name){
         System.out.println("Звонит абонент "+name);
     }
-    void receiveCall(String name, int phoneNumber){
+    void receiveCall(String name, long phoneNumber){
         System.out.println("Звонит абонент "+name+" номер "+phoneNumber);
     }
 }
